requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

asgiref==3.7.2
attrs==23.1.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
cloudinary==1.36.0
django==4.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
netifaces==0.11.0
packaging==23.2
Pillow==10.0.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
whitenoise==6.6.0
